+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   * Detect and notify of `301 Moved Permanently` responses.
   * Support editing `prevent_merge_without_jira_issue` when editing a project.
   * Support filtering projects by those that are hidden or are pending deletion.
+  * Support keyset pagination for the `/users` endpoint.
 
 ## Deprecations